feat: add twitter ui clone as example
This large PR brings in: 

- A near feature complete clone of the Twitter Tweet editor box
- Working mentions / suggestions (as inline-nodes for now). 
- Emoji support built on top of the impressive [emoji-mart](https://github.com/missive/emoji-mart) library.
- Enhanced link matching to mirror Twitter

**Motivation and Context**

Development on Remirror has stalled over the past month as a result of other projects taking priority. This PR should signal that it's an important project to me and has also given me time to become more familiar with the potential of a fully featured editor build with [prosemirror](https://prosemirror.net) as the foundation.

If possible I'd like to open it up to contributions from the community.

**How Has This Been Tested?**

Puppeteer has been added as an integration test framework. It was much simpler to work with when testing contenteditable, although the tests don't show up in coverage reports. 

Future work will need to address this.

I've started scoping out the potential of a `jest-remirror` library geared toward testing `remirror` editors.

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to change)

**Known bugs/issues**

- Only tested lazily in chrome so far (add e2e tests for other browsers)
- Adjacent emoji nodes (non-content nodes rendered via nodeViews) break the left and right arrow key and mess up the editor focus and cursor position. There's a test in `docs/examples/__tests__/ui-twitter.test.ts` which is currently skipped that can be run to see the fault in action
 - This is not something that can be solved via [`prosemirror-gapcursor`]()
 - Atlassian has [some code](https://bitbucket.org/atlassian/atlaskit-mk-2/src/8620f45319c3e66ea832f5db0ec3d179086676f7/packages/editor/editor-core/src/plugins/gap-cursor/?at=master) that looks pretty promising so once I understand more of it I'll implement a fix