fix: resolve rollup @emotion/core bundle issue
The babel configuration used for this project automatically injects
`jsx` from @emotion/core as a replacement for React.createElement.

This was causing problems with rollup which didn't see `@emotion/core`
as part of the dependencies and as a result would treat emotion as a
dependency and bundle it with the rest of the code.

Since emotion was bundled it meant that code consuming this library
could bundle emotion for node / browser and had to stick with the
browser bundle which this rollup build defaults to.

By adding @emotion/core as a dependency to any package which uses jsx,
rollup can now safely exclude @emotion/core from the bundle and the
downstream bug is fixed.