feat(react): enable automatic creation of manager
Manager is no longer a required prop for the `RemirrorProvider`. Now you
can also provide the extensions and presets to use as well as the
manager settings and your new manager will be created on the first
render.

**Please note:** Controlled editors can not make use of this shorthand
style since the manager is used in the creation of the initial state
value.