Prevent generating decorations without a class
The refractor highlighting produces two types of nodes:
* 'element' nodes indicate the type of the code snippet with the class
  ("token", as well as the specific thing such as "operator")
* 'text' nodes as children of the 'element' nodes that then contain the
  actual text of the snippet.

Depending on the language there can also be 'text' nodes outside of the context
of an 'element' node, for example in the TypeScript language a global variable
references would simply appear as a top-level 'text' node. As there are no classes
known for these nodes they should not get a decoration.

This commit introduces an option for the code-block extension to control this
behavior:
- { plainTextClassName: undefined } (the default): top-level 'text' nodes do not get decorated
- { plainTextClassName: 'some-class' }: top-level 'text' nodes get decorated with 'some-class'