Fix bottom value for multi-line annotations in positioner (#723)
The annotation positioner placed the bottom on the bottom of the first line. This meant that a popup placed on the bottom would overlay the rest of a multi-line annotation. Now, the positioner returns the bottom of the last line, allowing to place the popup below all content.

This commit contains also an addition to the storybook to actually see the problem (as there aren't any unit tests for positioners today).