feat(react-renderer): make composable (#881)
* Make react-renderer composable

Before, it was hard to add further node renderer to RemirrorRenderer. Users couldn't use the existing node renderer like TextHandler and CodeBlock because they weren't exported.
This commit simplifies composing a static renderer for the user's needs by exporting the renderers and the interface required to write custom node renderers.

As part of the commit, the complex renderer file is split into smaller - easier to understand&test - units.

* Add basic tests

* Support common node types by default

* Register remirror__react-renderer tests

* Add changeset

* Update .changeset/clean-rings-flash.md

Co-authored-by: Ifiok Jr. <ifiokotung@gmail.com>

Co-authored-by: Ifiok Jr. <ifiokotung@gmail.com>