Support static rendering of heading, iframe, link
Added support for some common node types and marks.

The iframe and link handler can be parameterized to change the visualization vs the node attributes in RemirrorJSON. For example, it allows to open links in a new tab or set the iframe to a specific width/height (independently of how this was entered in the editor).