fix(docs): enable embedded examples
The embedded examples would show the homepage instead of the code sample.

This commit sets the correct name for the example and adds structure to the examples (so further examples can be added).
Note: It would be better if structure could be added via sub directories (e.g. markdown/basic.tsx instead of markdown-basic/index.tsx) but I failed to make that work.