feat: refactor annotation backing store interface
Previously the annotations were always stored in something "MapLike", because that is a fairly natural
structure for the annotations extension. The yjs extension then got some magic so it could hook the needed
position transformations into it, and replace the "normal" Map with a Y.Map.

This commit tries to improve on that by extracting the concept of an "AnnotationStore", which allows to
clearer separate the concerns: The annotation extension doesn't need to know about position transformations
at all. On the other hand the that the Yjs extension has a Y.Map backing structure doesn't mean it should be
treated as a map but rather it should run certain changes in a "transaction", which is impossible to achieve
with the fine-grained `getMap` + transform position abstraction.

[1] From what I see this interface is "external", so for backwards-compatibility reasons I decided
    to retain the transformation magic also in the annotation extension, but the code is "dead" there
    and should actually be dropped when the support for the old options is removed.