feat: remove annotation support from yjs-extension (#1640)
We had initially added annotation support to the yjs-extension because we considered this to be required for collaborative editing with annotations. By now, this has proven to be wrong: sync'ing annotations outside of the Prosemirror document is bound to be unreliable (better approach: model annotations as marks).
Yet, having annotations hardwired in the yjs-extension gives a hard time to everyone wanting to use the yjs-extenion without annotations.

Requested a.o. by https://discord.com/channels/726035064831344711/745695521305526302/958036526443401217