feat: support marks to reference entities (#1648)
Before, the AnnotationExtension was the advised way to highlight content. Yet, this has proven to create lots of issues in combination with Collaborative Editing (see details in this [blog post](https://medium.com/collaborne-engineering/prosemirror-highlights-comments-20ce820149ed)).

The new EntityReferenceExtension allows to stores the highlights directly in the richtext as marks. The same approach can be used for other use cases like commenting richtext.

Co-authored-by: Ronny Roeller <ronny@collaborne.com>
Co-authored-by: ocavue <ocavue@gmail.com>
Co-authored-by: Will Hawker <w.hawker@hotmail.co.uk>